Don't pass a uint8_t* to sscanf for %x
For the %x format, sscanf expects an unsigned int*. The code previously
passed a uint8_t*, which caused it to write a zero byte to an undefined
memory location. In practice, this caused some other pointer to be
overwritten, which caused data reception to fail in some circumstances.

+++ src/AT/AtCmdLib.c
@@ -2786,7 +2786,7 @@ AtLib_UdpServerBulkDataTransfer (uint8_t cid, char *ipAddress, char *port,
 uint8_t
 AtLib_getCidFromAscii (uint8_t cidAscii)
 {
-  uint8_t cid = HOST_APP_INVALID_CID;
+  unsigned int cid = HOST_APP_INVALID_CID;
   uint8_t cidStr[] = " ";
   cidStr[0] = cidAscii;
